DDP-2340-consolidate-header-constants-into-class: consolidated the header constants
